% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMeasurementListToTibble.R
\name{convertMeasurementListToTibble}
\alias{convertMeasurementListToTibble}
\alias{convertmeasurementlisttotibble}
\title{convertMeasurementListToTibble}
\usage{
convertMeasurementListToTibble(df_data_list)
}
\arguments{
\item{df_data_list}{A list (list of data)}
}
\value{
A tibble with the data
}
\description{
Converts list with data to tibble
}
\details{
This function takes a list with oscilloscope readings saved
as yaml and coverts it to a tibble
}
\author{
Kai Budde-Sagert
}
