% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeasurements.R
\name{getMeasurements}
\alias{getMeasurements}
\alias{getmeasurements}
\title{getMeasurements}
\usage{
getMeasurements(input_file = NULL, input_directory = NULL)
}
\arguments{
\item{input_file}{A character (path to yaml file to be converted)}

\item{input_directory}{A character (path to directory with yaml files)}
}
\value{
A tibble with the data
}
\description{
Imports voltage measurement data from oscilloscope recordings
}
\details{
This function reads yaml files (from a directory) or a zip file
and imports the data into a tibble
}
\author{
Kai Budde
}
