% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaveforms.R
\name{getWaveforms}
\alias{getWaveforms}
\alias{getwaveforms}
\title{getWaveforms}
\usage{
getWaveforms(
  input_file = NULL,
  input_directory = NULL,
  reduce_resolution_of_input_data = TRUE
)
}
\arguments{
\item{input_file}{A character (path to yaml file to be converted)}

\item{input_directory}{A character (path to directory with yaml files)}

\item{reduce_resolution_of_input_data}{A Boolean (showing if fewer point
are to be imported)}
}
\value{
A tibble with the data
}
\description{
Imports waveform data from oscilloscope recordings
}
\details{
This function reads yaml files (from a directory) or a zip file
and imports the waveform data into a tibble
}
\author{
Kai Budde-Sagert
}
