% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasurements.R
\name{plotMeasurements}
\alias{plotMeasurements}
\alias{plotmeasurements}
\title{plotMeasurements}
\usage{
plotMeasurements(
  input_data = NULL,
  output_dir = NULL,
  vavg_min = NULL,
  vavg_max = NULL,
  vpp_min = NULL,
  vpp_max = NULL,
  plot_title = NULL
)
}
\arguments{
\item{input_data}{A tibble (measurement data)}

\item{output_dir}{A character (path to output directory)}

\item{vavg_min}{A number (min value of Vavg to be plotted)}

\item{vavg_max}{A number (max value of Vavg to be plotted)}

\item{vpp_min}{A number (min value of Vpp to be plotted)}

\item{vpp_max}{A number (max value of Vpp to be plotted)}

\item{plot_title}{A character (title of the plot)}
}
\value{
0
}
\description{
Plots imported measurment data
}
\details{
This functions takes recorded oscilloscope data (Vavg and Vpp)
(obtained through getMeasurements.R) and plots it
}
\author{
Kai Budde-Sagert
}
