% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaveforms_simple.R
\name{plotWaveforms_simple}
\alias{plotWaveforms_simple}
\alias{plotwaveformssimple,}
\alias{plotwaveforms_simple}
\title{plotWaveforms_simple}
\usage{
plotWaveforms_simple(
  input_data = NULL,
  output_dir = NULL,
  voltage_limits_of_plot = NULL,
  plot_title = NULL,
  print_width = 19.2,
  print_height = 10.8
)
}
\arguments{
\item{input_data}{A tibble (measurement data)}

\item{output_dir}{A character (path to output directory)}

\item{voltage_limits_of_plot}{A number (value of min (negativ of the value)
and max values to be shown on y axis)}

\item{plot_title}{A character (title of the plot)}

\item{print_width}{A number (width in cm of the saved plot)}

\item{print_height}{A number (height in cm of the saved plot)}
}
\value{
0
}
\description{
Plots imported waveform data without any calculations
}
\details{
This functions takes recorded oscilloscope data (waveforms)
(obtained through getWaveforms.R) and plots it
}
\author{
Kai Budde-Sagert
}
