% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaveforms.R
\name{plotWaveforms}
\alias{plotWaveforms}
\alias{plotwaveforms}
\title{plotWaveforms}
\usage{
plotWaveforms(
  input_data = NULL,
  output_dir = NULL,
  show_time_in_us = FALSE,
  channel_function_generator = "CHAN1",
  channel_stimulation_pulse = "CHAN2",
  plot_waveforms = "all",
  voltage_limits_of_plot = NULL,
  filter_stim_off = TRUE,
  epsilon_for_filtering = 1.5,
  plot_title = NULL
)
}
\arguments{
\item{input_data}{A tibble (measurement data)}

\item{output_dir}{A character (path to output directory)}

\item{show_time_in_us}{A boolean (indicates whether time is to be shown
in us)}

\item{channel_function_generator}{A character (name of the channel with
the function generator pulse (for GATE TTL))}

\item{channel_stimulation_pulse}{A character (name of the channel with
the stimulation pulse)}

\item{plot_waveforms}{A character (shows if all, none, or one waveform shall be plotted)}

\item{voltage_limits_of_plot}{A number (value of min (negativ of the value)
and max values to be shown on y axis)}

\item{filter_stim_off}{A boolean (indicates whether measurements while
stim==off are to be filtered)}

\item{epsilon_for_filtering}{A number (min. distance for mean from 0 needed for filtering out the signal generator)}

\item{plot_title}{A character (title of the plot)}
}
\value{
0
}
\description{
Plots imported waveform data
}
\details{
This functions takes recorded oscilloscope data (waveforms)
(obtained through getWaveforms.R) and plots it
}
\author{
Kai Budde
}
